+++ uefi/src/data_types/owned_strs.rs
@@ -3,6 +3,7 @@ use super::strs::{CStr16, FromSliceWithNulError};
 use crate::data_types::strs::EqStrUntilNul;
 use crate::data_types::UnalignedSlice;
 use crate::polyfill::vec_into_raw_parts;
+use alloc::borrow::{Borrow, ToOwned};
 use alloc::vec::Vec;
 use core::fmt;
 use core::ops;
@@ -133,6 +134,20 @@ impl AsRef<CStr16> for CString16 {
     }
 }
 
+impl Borrow<CStr16> for CString16 {
+    fn borrow(&self) -> &CStr16 {
+        self
+    }
+}
+
+impl ToOwned for CStr16 {
+    type Owned = CString16;
+
+    fn to_owned(&self) -> CString16 {
+        CString16(self.as_slice_with_nul().to_vec())
+    }
+}
+
 impl fmt::Display for CString16 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         self.as_ref().fmt(f)
@@ -155,6 +170,7 @@ impl<StrType: AsRef<str> + ?Sized> EqStrUntilNul<StrType> for CString16 {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::cstr16;
     use alloc::string::String;
     use alloc::vec;
 
@@ -223,5 +239,22 @@ mod tests {
         // now other direction
         assert!(String::from("test").eq_str_until_nul(&input));
         assert!("test".eq_str_until_nul(&input));
+    }
+
+    /// Test the `Borrow` and `ToOwned` impls.
+    #[test]
+    fn test_borrow_and_to_owned() {
+        let s1: &CStr16 = cstr16!("ab");
+        let owned: CString16 = s1.to_owned();
+        let s2: &CStr16 = owned.borrow();
+        assert_eq!(s1, s2);
+        assert_eq!(
+            owned.0,
+            [
+                Char16::try_from('a').unwrap(),
+                Char16::try_from('b').unwrap(),
+                NUL_16
+            ]
+        );
     }
 }